Encapsulate Data Store Summarization and Context
